Adding Hacking with React (#12)
A friend recommended me this guide to start react and I'm about halfway through the book. It is a really cool and hands-on based approach that takes you step-by-step through creating a bigger application, great for the kind of developer that much like me needs to get their toes wet before gaining any kind of motivation to read a slightly more technical/documentation heavy text.